Refactor: add validation and change data type of score
java 에서 실수형 사용시 BigDecimal 사용

+++ src/main/java/com/devthink/devthink_server/dto/ReviewResponseDto.java
@@ -3,6 +3,8 @@ package com.devthink.devthink_server.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class ReviewResponseDto {
@@ -15,10 +17,10 @@ public class ReviewResponseDto {
 
     private String content;
 
-    private Float score;
+    private BigDecimal score;
 
     @Builder
-    public ReviewResponseDto(Long id, Long userId,Integer bookIsbn,String content,Float score){
+    public ReviewResponseDto(Long id, Long userId, Integer bookIsbn, String content, BigDecimal score){
         this.id = id;
         this.userId = userId;
         this.bookIsbn = bookIsbn;
